refactor: move aac to own package
